requirements.txt
rpy2==3.3.2
discord.py==1.0.0